requirements.txt
flask
Flask
openpyxl
pandas
pyinstaller
pyserial
pyserial-asyncio
python-tkdnd
pystray